+++ 26.2Fabric/src/main/java/com/abyssredemption/abstool/fabric/AbsToolFabricClient.java
@@ -21,6 +21,7 @@ public final class AbsToolFabricClient implements ClientModInitializer {
     public void onInitializeClient() {
         com.abyssredemption.abstool.client.furnace.FurnaceNetwork.canSend = net.fabricmc.fabric.api.client.networking.v1.ClientPlayNetworking::canSend;
         AbsToolClient.init(FabricLoader.getInstance().getConfigDir());
+        ServerStatsNetworking.init();
         com.abyssredemption.abstool.fabric.compat.schematic.SchematicBootstrap.init();
         if (FabricLoader.getInstance().isModLoaded("tweakermore")) {
             com.abyssredemption.abstool.client.vault.config.QuickSettings.provider =