+++ common/src/main/java/com/abyssredemption/abstool/client/AbsToolClient.java
@@ -22,7 +22,6 @@ public final class AbsToolClient {
     }
 
     public static void tickShortcut(Minecraft client) {
-        ServerStatsNavigation.tick();
         com.abyssredemption.abstool.client.furnace.FurnaceTracker.INSTANCE.tick(client);
         com.abyssredemption.abstool.client.elytra.ElytraLaunch.guard(client);
         boolean focused = client.isWindowActive();